# backend/docs/swagger.yaml
basePath: /
definitions:
  internal_httpserver_handlers.ErrorResponse:
    properties:
      error:
        type: string
      ok:
        type: boolean
    type: object
  internal_httpserver_handlers.GenericObject:
    additionalProperties: {}
    type: object
  internal_httpserver_handlers.MpAuthLoginResponse:
    properties:
      expiresAt:
        type: string
      ok:
        type: boolean
      token:
        type: string
      user:
        $ref: '#/definitions/internal_httpserver_handlers.MpAuthUser'
    type: object
  internal_httpserver_handlers.MpAuthUser:
    properties:
      id:
        type: integer
      openid:
        type: string
    type: object
...
  internal_httpserver_handlers.OkResponse:
    properties:
      ok:
        type: boolean
    type: object
...
  internal_httpserver_handlers.mpAuthBindDeviceRequest:
    properties:
      device_id:
        type: string
    type: object
  internal_httpserver_handlers.mpAuthLoginRequest:
    properties:
      code:
        type: string
    type: object
  internal_httpserver_handlers.mpAuthUpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      nick_name:
        type: string
    type: object
...
info:
  contact: {}
  description: 后端基于 Gin 的 HTTP API 文档（swaggo/swag 自动生成）。
  title: F1InkDashboard Backend API
  version: 1.0.0
paths:
...
  /api/v1/mp/auth/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 头像图片文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 上传用户头像
      tags:
      - MiniProgramAuth
  /api/v1/mp/auth/bind_device:
    post:
      consumes:
      - application/json
      description: 绑定前要求设备已调用 /api/v1/device/boot 上报（后端会校验 boot report 存在）。
      parameters:
      - description: 绑定请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpserver_handlers.mpAuthBindDeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 绑定设备
      tags:
      - MiniProgramAuth
  /api/v1/mp/auth/login:
    post:
      consumes:
      - application/json
      description: 使用 wx.login 获取的 code 换取 openid 并返回后端 token。
      parameters:
      - description: 登录请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpserver_handlers.mpAuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.MpAuthLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      summary: 小程序登录
      tags:
      - MiniProgramAuth
  /api/v1/mp/auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.OkResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 登出
      tags:
      - MiniProgramAuth
  /api/v1/mp/auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.GenericObject'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - MiniProgramAuth
...
  /api/v1/mp/auth/profile:
    post:
      consumes:
      - application/json
      description: nick_name/avatar_url 至少传一个非空。
      parameters:
      - description: 更新内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_httpserver_handlers.mpAuthUpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.OkResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/internal_httpserver_handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新用户资料
      tags:
      - MiniProgramAuth
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Authorization: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  TokenQuery:
    description: 简单鉴权 token（NEWS_INGEST_TOKEN / OPENF1_INGEST_TOKEN / WECHATPAY_API_TOKEN
      等）
    in: query
    name: token
    type: apiKey
swagger: "2.0"